requirements.txt
pybaseball>=2.2.0
tweepy>=4.14.0
matplotlib>=3.7.0
Pillow>=10.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
